add SSO library

+++ V1.0/views/layouts/AdminLTE/base.php
@@ -2,7 +2,10 @@
     <div class="col-xs-12">
         <div class="box box-info">
             <div class="box-header with-border">
-                <?php echo heading(anchor($base . '/create/', '<i class="fa fa-plus"></i> Create New'), 3, 'class="box-title"') ?>
+                <?php 
+                if(in_array('c',$this->sso_new->curr_access))
+                    echo heading(anchor($base . '/create/', '<i class="fa fa-plus"></i> Create New'), 3, 'class="box-title"') 
+                ?>
                 <div class="box-tools pull-right">
                     <?php echo form_open('#', ['method' => 'GET']) ?>
                     <div class="input-group input-group-sm">
@@ -31,8 +34,12 @@
                                     $key = array_keys(object_to_array($row));
                                     $row->{$key[1]} = $row->{next($key)};
                                     $row->{$key[1]} .= sprintf('<div class="small action" id="qe-%s">', $row->$id);
-                                    $row->{$key[1]} .= anchor($base . '/update/' . $row->$id, '<i class="fa fa-edit"></i>' . nbs() . 'Edit', 'title="Edit data"') . nbs(2) . '<span class="text-muted small">|</span>' . nbs(2);
-                                    $row->{$key[1]} .= anchor('#', '<i class="fa fa-trash"></i>' . nbs() . 'Delete', 'class="text-danger" title="Delete user" data-toggle="modal" data-target="#confirm-delete" data-href="' . base_url($base . '/delete/' . $row->$id) . '"');
+                                    if(in_array('u',$this->sso_new->curr_access))
+                                        $row->{$key[1]} .= anchor($base . '/update/' . $row->$id, '<i class="fa fa-edit"></i>' . nbs() . 'Edit', 'title="Edit data"') . nbs(2) . '<span class="text-muted small"></span>' . nbs(2);
+                                    
+                                    if(in_array('d',$this->sso_new->curr_access))
+                                        $row->{$key[1]} .= anchor('#', '<i class="fa fa-trash"></i>' . nbs() . 'Delete', 'class="text-danger" title="Delete user" data-toggle="modal" data-target="#confirm-delete" data-href="' . base_url($base . '/delete/' . $row->$id) . '"');
+                                    
                                     $row->{$key[1]} .= '</div>';
                                     if ($show_pk === FALSE)
                                         unset($row->$id);
